manage listings page, user authorization

+++ app/Http/Controllers/ListingController.php
@@ -69,6 +69,11 @@ class ListingController extends Controller
     //Update Listing Data
     public function update(Request $request, Listing $listing)
     {
+        //Make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         /* dd($request->all()); */
         $formFields = $request->validate([
             'title' => 'required',
@@ -92,7 +97,23 @@ class ListingController extends Controller
     //Delete Listing
     public function destroy(Listing $listing)
     {
+        //Make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully!');
+    }
+
+    //Manage Listings
+    public function manage()
+    {
+        return view('listings.manage', [
+            'listings' => auth()
+                ->user()
+                ->listings()
+                ->get(),
+        ]);
     }
 }
